refactor2.  util경로설정
