+++ public/js/view_more_manager.js
@@ -1,4 +1,4 @@
-import { _ } from './util.js';
+import { _ } from './util/util.js';
 
 export class ViewMoreManager {
   constructor(titleArr, descArr, imgUrlArr, value) {
